# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateInviteRequest": {
        "properties": {
          "role": {
            "description": "`\"node\"` (default) or `\"admin\"`.",
            "type": "string"
          },
          "ttl_seconds": {
            "description": "Lifetime in seconds. Capped at 30 days; 0 falls back to one hour.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "InviteResponse": {
        "properties": {
          "cluster": {
            "description": "Cluster name this invite admits the new node into.",
            "type": "string"
          },
          "expires_in": {
            "description": "Number of seconds remaining before the token expires.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "role": {
            "description": "Role the new node will receive (`\"node\"` or `\"admin\"`).",
            "type": "string"
          },
          "token": {
            "description": "Opaque invite token (CBOR + Ed25519 signature, base64-url encoded).",
            "type": "string"
          },
          "url": {
            "description": "Ready-to-paste URL for `mlsh adopt`.",
            "type": "string"
          }
        },
        "required": [
          "token",
          "url",
          "cluster",
          "role",
          "expires_in"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API for managing the MLSH overlay network.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "MLSH Control API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/invites": {
      "post": {
        "operationId": "create_invite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteResponse"
                }
              }
            },
            "description": "Invite created"
          },
          "400": {
            "description": "Invalid role or empty key"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Sponsor key unreadable or signing failed"
          }
        },
        "summary": "Create a new sponsor-signed invite. The token is stateless: the response\nis the *only* time the server sees it. Lose it and you must re-issue.",
        "tags": [
          "Invites"
        ]
      }
    },
...
  }
}